Update struct data structure with tight packing

+++ src/Common.sol
@@ -3,14 +3,14 @@ pragma solidity 0.8.13;
 
 library Common {
     struct GlobalState {
-        uint256 lastUpdate;
-        uint256 lastSupply;
+        uint32 lastUpdate;
+        uint224 lastSupply;
         uint256 rewards;
     }
 
     struct UserState {
-        uint256 lastUpdate;
-        uint256 lastBalance;
+        uint32 lastUpdate;
+        uint224 lastBalance;
         uint256 rewards;
     }
 }
